MB-67520: Fix many narrowing warnings in ep-engine
Note for changes in visitDepth (stat generation) it is safe to remove
the -1 special case. visitDepth will be called as there is no
HashTable with 0 buckets and as such a min depth will be calculated.

Reviewed-on: https://review.couchbase.org/c/kv_engine/+/230773
Tested-by: Jim Walker <jim@couchbase.com>

+++ engines/ep/src/bloomfilter.cc
@@ -13,6 +13,7 @@
 
 #include "bloomfilter.h"
 
+#include <gsl/gsl-lite.hpp>
 #include <platform/murmurhash3.h>
 
 #include <cmath>
@@ -48,7 +49,8 @@ size_t BloomFilter::estimateNoOfHashes(size_t key_count) {
 uint64_t BloomFilter::hashDocKey(const DocKeyView& key, uint32_t iteration) {
     uint64_t result = 0;
     auto hashable = key.getIdAndKey();
-    uint32_t seed = iteration + (uint32_t(hashable.first) * noOfHashes);
+    uint32_t seed = iteration + (uint32_t(hashable.first) *
+                                 gsl::narrow_cast<uint32_t>(noOfHashes));
     MURMURHASH_3(hashable.second.data(), hashable.second.size(), seed, &result);
     return result;
 }
